requirement.txt
asgiref==3.4.1
attrs==21.2.0
black==21.12b0
Cerberus==1.3.4
click==8.0.3
Django==3.2.10
django-debug-toolbar==3.2.4
django-extensions==3.1.5
django-filter==21.1
django-security==0.12.0
djangorestframework==3.13.1
flake8==4.0.1
gunicorn==20.1.0
importlib-metadata==4.10.0
iniconfig==1.1.1
Markdown==3.3.6
mccabe==0.6.1
mypy==0.920
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.0
pluggy==1.0.0
ptvsd==4.3.2
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.11.2
pyparsing==3.0.6
pytest==6.2.5
pytest-mock==3.6.1
python-dateutil==2.8.0
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.3
typed-ast==1.5.1
typing_extensions==4.0.1
ua-parser==0.10.0
Werkzeug==2.0.2
zipp==3.6.0
